requirements.txt
snowflake-snowpark-python>=1.0.0
click>=8.0.0
toml>=0.10.0
pytest>=7.0.0
